Replace ImageView trait with IntoXBuffer, support image::DynamicImage
Replace ImageView trait which provides a method to convert an image into RGB
buffer with IntoXBuffer trait which has a method for converting into image
format defined by the X display server.

Significant benefit of this approach is that if an image is not in Rgb8 format
format, we avoid intermediate conversion to that format.  With ImageView the
flow for Rgb16 was to first convert it to Rgb8 and then the crate would convert
that to Rgba8.  With IntoXBuffer, the image is converted directly to Rgba8.

With this change also add support for image::DynamicImage type rather than
limiting support to just RgbImage.